+++ doc/CHANGELOG.md
@@ -29,16 +29,21 @@ See doc/COPYRIGHT.rdoc for more details.
 
 # Changelog
 
+* `#1281` I18n.js Not working correctly. Always returns English Translations
+* `#1771` Fixed bug: Refactor Types Project Settings into new Tab
 * `#1880` Re-introduce at-time scope
 * `#1881` Re-introduce project planning comparison in controller
 * `#1883` Extend at-time scope for status comparison
 * `#1884` Make status values available over API
 * `#1994` Integrational tests for work packages at_time (API)
+* `#2158` Work Package General Setting
 * `#2173` Adapt client-side to new server behavior
+* `#2307` Change icon of home button in header from OpenProjct icon to house icon
+* `#2310` Add proper indices to work_package
+* `#2319` Add a request-store to avoid redundant calls
 
 ## 3.0.0pre20
 
-* `#1281` I18n.js Not working correctly. Always returns English Translations
 * `#1560` WorkPackage/update does not retain some fields when validations fail
 * `#1771` Refactor Types Project Settings into new Tab
 * `#1878` Project Plan Comparison(server-side implementation): api/v2 can now resolve historical data for work_packages
@@ -51,7 +56,6 @@ See doc/COPYRIGHT.rdoc for more details.
 * `#2246` Migrate issue categories controller tests
 * `#2262` Migrate issue statuses controller tests
 * `#2267` Rename view issue hooks
-* `#2307` Change icon of home button in header from OpenProjct icon to house icon
 
 ## 3.0.0pre19