+++ doc/CHANGELOG.md
@@ -30,6 +30,7 @@ See doc/COPYRIGHT.rdoc for more details.
 # Changelog
 
 * `#1281` I18n.js Not working correctly. Always returns English Translations
+* `#1758` Migrate functional-tests for issues into specs for work package
 * `#1771` Fixed bug: Refactor Types Project Settings into new Tab
 * `#2306` Migrate issues controller tests
 * `#2307` Change icon of home button in header from OpenProjct icon to house icon