+++ doc/CHANGELOG.md
@@ -32,8 +32,10 @@ See doc/COPYRIGHT.rdoc for more details.
 * `#1281` I18n.js Not working correctly. Always returns English Translations
 * `#1758` Migrate functional-tests for issues into specs for work package
 * `#1771` Fixed bug: Refactor Types Project Settings into new Tab
+* `#2158` Work Package General Setting
 * `#2306` Migrate issues controller tests
 * `#2307` Change icon of home button in header from OpenProjct icon to house icon
+* `#2310` Add proper indices to work_package
 * `#2319` Add a request-store to avoid redundant calls
 
 ## 3.0.0pre20